Move CSV/JSON generation to test helpers
